refactor(computer-use): extract TOCTOU target re-verification from cua-driver-backend.ts
